requirements.txt
bcrypt==3.2.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
Jinja2==2.11.2
numpy==1.18.5
pdfkit==0.6.1
Pillow==7.2.0
requests==2.24.0
requests-oauthlib==1.3.0
SQLAlchemy==1.3.22
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
Flask-Mail==0.9.1
psycopg2-binary==2.8.6
gunicorn==20.0.4
boto3==1.17.25
Flask-Migrate==2.7.0